+++ db/fdb_bend_sql.c
@@ -983,7 +983,7 @@ struct sqlclntstate *fdb_svc_trans_get(char *tid, int isuuid)
     struct sqlclntstate *clnt;
     int rc = 0;
 
-    int                  deadline =  0;
+    int64_t deadline = 0;
     int                  wait = bdb_attr_get(thedb->bdb_attr, 
             BDB_ATTR_TIMEOUT_FDB_TRANS_SYNC);
 
@@ -999,8 +999,8 @@ struct sqlclntstate *fdb_svc_trans_get(char *tid, int isuuid)
             rc = osql_chkboard_get_clnt(*(unsigned long long *)tid, &clnt);
         if (rc && rc == -1) {
             if (deadline && comdb2_time_epochms() > deadline) {
-                logmsg(LOGMSG_ERROR, "%s: timeout waiting for transaction %d waited %u\n",
-                        __func__, deadline, wait);
+                logmsg(LOGMSG_ERROR, "%s: timeout waiting for transaction %" PRId64 " waited %u\n", __func__, deadline,
+                       wait);
 
                 break;
             }
