+++ schemachange/schemachange.c
@@ -415,7 +415,7 @@ void delay_if_sc_resuming(struct ireq *iq)
 
     int diff;
     int printerr = 0;
-    int start_time = comdb2_time_epochms();
+    int64_t start_time = comdb2_time_epochms();
     while (gbl_sc_resume_start > 0) {
         if ((diff = comdb2_time_epochms() - start_time) > 300 && !printerr) {
             logmsg(LOGMSG_WARN, "Delaying since gbl_sc_resume_start has not "
